# Conflicts:
#	packages/examples/cvat/exchange-oracle/poetry.lock
#	packages/examples/cvat/exchange-oracle/pyproject.toml
#	packages/examples/cvat/exchange-oracle/src/core/config.py
#	packages/examples/cvat/exchange-oracle/src/core/manifest.py
#	packages/examples/cvat/exchange-oracle/src/endpoints/__init__.py
#	packages/examples/cvat/exchange-oracle/src/endpoints/exchange.py
#	packages/examples/cvat/exchange-oracle/src/schemas/__init__.py
#	packages/examples/cvat/exchange-oracle/src/schemas/cvat.py
#	packages/examples/cvat/exchange-oracle/src/schemas/exchange.py
#	packages/examples/cvat/exchange-oracle/src/services/cloud/types.py
#	packages/examples/cvat/exchange-oracle/src/services/cvat.py
#	packages/examples/cvat/exchange-oracle/src/services/exchange.py
#	packages/examples/cvat/exchange-oracle/tests/api/test_exchange_api.py
#	packages/examples/cvat/exchange-oracle/tests/integration/cron/state_trackers/test_track_completed_escrows.py
#	packages/examples/cvat/exchange-oracle/tests/integration/cron/state_trackers/test_track_task_creation.py
#	packages/examples/cvat/exchange-oracle/tests/integration/cron/test_process_recording_oracle_webhooks.py
#	packages/examples/cvat/exchange-oracle/tests/integration/services/test_exchange.py
#	packages/examples/cvat/exchange-oracle/tests/integration/services/test_webhook.py
#	packages/examples/cvat/recording-oracle/poetry.lock
#	packages/examples/cvat/recording-oracle/src/.env.template
#	packages/examples/cvat/recording-oracle/src/core/manifest.py
#	packages/examples/cvat/recording-oracle/src/core/validation_results.py
#	packages/examples/cvat/recording-oracle/src/endpoints/__init__.py
#	packages/examples/cvat/recording-oracle/src/handlers/validation.py
#	packages/examples/cvat/recording-oracle/src/schemas/__init__.py
#	packages/examples/cvat/recording-oracle/src/services/cloud/types.py
#	packages/examples/cvat/recording-oracle/tests/integration/services/test_webhook_service.py

+++ packages/apps/reputation-oracle/server/src/modules/qualification/qualification.service.ts
@@ -95,6 +95,13 @@ export class QualificationService {
       );
     }
 
+    if (qualificationEntity.userQualifications.length > 0) {
+      throw new ControlledError(
+        ErrorQualification.CannotDeleteAssignedQualification,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return this.qualificationRepository.deleteOne(qualificationEntity);
   }
